69 create handler of entering name for unregistered users(closes #69) (#73)
* Display message sending time (closes #70)

* Minor refactoring

* Added Notification

May be used for any purpose. For now it is used to notify user the same way console.log does, but it may be changed in the future to serve other purpose

* Enter Name modal

It is used to enter chat as guest.
At the moment any name but blank is accepted.  If user is registered he will be logged in automatically. It other cases, user will have a new name and new session on every connection

* added handling for some minor error cases

* BE part of guest log in

- contains small bug fixes
- also changed sended to sent :alien:

Co-authored-by: Filipp-Artur Pljassunov <33956131+FilipsP@users.noreply.github.com>